Initial Commit of the work that I've done.
Right now the "plane" flies around a circle around the "airfield". We have a runway object, that generates two end points that can be used to figure out when the plane should turn in if the landing is triggered. I don't think that the current method of designating the "turn in point" is that solid yet. Need to think about it more.